Issue #154 - Move functions on the proper package
